+++ src/main/java/com/rest/documentManager/service/AttachmentService.java
@@ -6,5 +6,5 @@ import org.springframework.web.multipart.MultipartFile;
 public interface AttachmentService {
     Attachment saveAttachment(MultipartFile file) throws Exception;
 
-    Attachment getAttachment(Integer id) throws Exception;
+    Attachment getAttachment(Long id) throws Exception;
 }
